requirements.txt
amqp==5.3.1
asgiref==3.7.2
async-timeout==4.0.3
backports.zoneinfo==0.2.1
billiard==3.6.4.0
cached-property==1.5.2
celery==5.2.7
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==3.2.25
django-filter==23.5
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.8
Faker==18.13.0
gunicorn==23.0.0
importlib-metadata==6.7.0
inflection==0.5.1
kombu==5.2.4
Markdown==3.4.4
packaging==24.0
prompt_toolkit==3.0.48
psycopg2==2.9.9
PyJWT==2.8.0
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.1
redis==5.0.8
six==1.17.0
sqlparse==0.4.4
typing_extensions==4.7.1
uritemplate==4.1.1
vine==5.1.0
wcwidth==0.2.13
zipp==3.15.0
